Detect GOPATH if not specified in $GOPATH
It is no longer required to have the GOPATH environment variable set -
it now defaults to ~/go, but irrespective of whether the value is the
default or not it can be found in the standard output of `go env
GOPATH`.

This commit uses that value if the GOPATH environment variable is not
set.