Fix bug in handling on anon structs which embed other structs
If the anon struct embeds another struct _and_ contains fields of its own,
there were cases where the generated code would ignore the anon struct's own
fields.

This happened because when checking if the anon struct implements a marshaling
interface, the reflect package returns true if the anon struct's _embedded
struct_ implements that interface. But using that interface means that the
other fields in the anon struct are ignored.